feat(): move to new models

+++ src/main/java/com/beandon/backend/pojo/stocks/AccountOverview.java
@@ -8,6 +8,6 @@ import lombok.Data;
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class AccountOverview {
-    private int amount;
-    private double percentChange;
+    private float amount;
+    private float percentChange;
 }
